Refactoring list records and list records in view components, extracting common list logic into its own module.
